+++ besu/src/main/java/org/hyperledger/besu/cli/subcommands/operator/RestoreState.java
@@ -35,6 +35,7 @@ import org.hyperledger.besu.ethereum.rlp.RLPInput;
 import org.hyperledger.besu.ethereum.trie.Node;
 import org.hyperledger.besu.ethereum.trie.PersistVisitor;
 import org.hyperledger.besu.ethereum.trie.RestoreVisitor;
+import org.hyperledger.besu.ethereum.worldstate.DefaultWorldStateArchive;
 import org.hyperledger.besu.ethereum.worldstate.StateTrieAccountValue;
 import org.hyperledger.besu.ethereum.worldstate.WorldStateStorage;
 
@@ -251,7 +252,8 @@ public class RestoreState implements Runnable {
       updater.commit();
     }
     final WorldStateStorage worldStateStorage =
-        besuController.getProtocolContext().getWorldStateArchive().getWorldStateStorage();
+        ((DefaultWorldStateArchive) besuController.getProtocolContext().getWorldStateArchive())
+            .getWorldStateStorage();
     updater = worldStateStorage.updater();
   }
