+++ besu/src/main/java/org/hyperledger/besu/controller/BesuControllerBuilder.java
@@ -52,10 +52,13 @@ import org.hyperledger.besu.ethereum.eth.transactions.TransactionPoolFactory;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSchedule;
 import org.hyperledger.besu.ethereum.p2p.config.SubProtocolConfiguration;
 import org.hyperledger.besu.ethereum.storage.StorageProvider;
+import org.hyperledger.besu.ethereum.worldstate.DefaultWorldStateArchive;
 import org.hyperledger.besu.ethereum.worldstate.MarkSweepPruner;
 import org.hyperledger.besu.ethereum.worldstate.Pruner;
 import org.hyperledger.besu.ethereum.worldstate.PrunerConfiguration;
 import org.hyperledger.besu.ethereum.worldstate.WorldStateArchive;
+import org.hyperledger.besu.ethereum.worldstate.WorldStatePreimageStorage;
+import org.hyperledger.besu.ethereum.worldstate.WorldStateStorage;
 import org.hyperledger.besu.metrics.ObservableMetricsSystem;
 
 import java.io.Closeable;
@@ -213,9 +216,12 @@ public abstract class BesuControllerBuilder {
 
     final ProtocolSchedule protocolSchedule = createProtocolSchedule();
     final GenesisState genesisState = GenesisState.fromConfig(genesisConfig, protocolSchedule);
+    final WorldStateStorage worldStateStorage = storageProvider.createWorldStateStorage();
+    final WorldStateArchive worldStateArchive = createWorldStateArchive(worldStateStorage);
     final ProtocolContext protocolContext =
         ProtocolContext.init(
             storageProvider,
+            worldStateArchive,
             genesisState,
             protocolSchedule,
             metricsSystem,
@@ -238,7 +244,7 @@ public abstract class BesuControllerBuilder {
             Optional.of(
                 new Pruner(
                     new MarkSweepPruner(
-                        protocolContext.getWorldStateArchive().getWorldStateStorage(),
+                        ((DefaultWorldStateArchive) worldStateArchive).getWorldStateStorage(),
                         blockchain,
                         storageProvider.createPruningStorage(),
                         metricsSystem),
@@ -297,7 +303,7 @@ public abstract class BesuControllerBuilder {
             syncConfig,
             protocolSchedule,
             protocolContext,
-            protocolContext.getWorldStateArchive().getWorldStateStorage(),
+            worldStateStorage,
             ethProtocolManager.getBlockBroadcaster(),
             maybePruner,
             ethProtocolManager.ethContext(),
@@ -402,6 +408,12 @@ public abstract class BesuControllerBuilder {
         fastSyncEnabled,
         scheduler,
         genesisConfig.getForks());
+  }
+
+  public WorldStateArchive createWorldStateArchive(final WorldStateStorage worldStateStorage) {
+    final WorldStatePreimageStorage preimageStorage =
+        storageProvider.createWorldStatePreimageStorage();
+    return new DefaultWorldStateArchive(worldStateStorage, preimageStorage);
   }
 
   private List<PeerValidator> createPeerValidators(final ProtocolSchedule protocolSchedule) {
