+++ ethereum/core/build.gradle
@@ -48,6 +48,9 @@ dependencies {
   implementation 'org.apache.tuweni:tuweni-units'
   implementation 'org.hyperledger.besu:altbn128'
   implementation 'org.hyperledger.besu:bls12-381'
+  implementation 'org.immutables:value-annotations'
+
+  annotationProcessor 'org.immutables:value'
 
   runtimeOnly 'org.apache.logging.log4j:log4j-core'
