+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/ProtocolContext.java
@@ -22,8 +22,6 @@ import org.hyperledger.besu.ethereum.chain.MutableBlockchain;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSchedule;
 import org.hyperledger.besu.ethereum.storage.StorageProvider;
 import org.hyperledger.besu.ethereum.worldstate.WorldStateArchive;
-import org.hyperledger.besu.ethereum.worldstate.WorldStatePreimageStorage;
-import org.hyperledger.besu.ethereum.worldstate.WorldStateStorage;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
 
 import java.util.function.BiFunction;
@@ -49,6 +47,7 @@ public class ProtocolContext {
 
   public static ProtocolContext init(
       final StorageProvider storageProvider,
+      final WorldStateArchive worldStateArchive,
       final GenesisState genesisState,
       final ProtocolSchedule protocolSchedule,
       final MetricsSystem metricsSystem,
@@ -56,17 +55,14 @@ public class ProtocolContext {
       final long reorgLoggingThreshold) {
     final BlockchainStorage blockchainStorage =
         storageProvider.createBlockchainStorage(protocolSchedule);
-    final WorldStateStorage worldStateStorage = storageProvider.createWorldStateStorage();
-    final WorldStatePreimageStorage preimageStorage =
-        storageProvider.createWorldStatePreimageStorage();
 
     final MutableBlockchain blockchain =
         DefaultBlockchain.createMutable(
             genesisState.getBlock(), blockchainStorage, metricsSystem, reorgLoggingThreshold);
 
-    final WorldStateArchive worldStateArchive =
-        new WorldStateArchive(worldStateStorage, preimageStorage);
-    genesisState.writeStateTo(worldStateArchive.getMutable());
+    if (blockchain.getChainHeadBlockNumber() < 1) {
+      genesisState.writeStateTo(worldStateArchive.getMutable());
+    }
 
     return new ProtocolContext(
         blockchain,
