+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/mainnet/MainnetTransactionProcessor.java
@@ -17,7 +17,7 @@ package org.hyperledger.besu.ethereum.mainnet;
 import org.hyperledger.besu.ethereum.chain.Blockchain;
 import org.hyperledger.besu.ethereum.core.Account;
 import org.hyperledger.besu.ethereum.core.Address;
-import org.hyperledger.besu.ethereum.core.DefaultEvmAccount;
+import org.hyperledger.besu.ethereum.core.EvmAccount;
 import org.hyperledger.besu.ethereum.core.Gas;
 import org.hyperledger.besu.ethereum.core.Log;
 import org.hyperledger.besu.ethereum.core.MutableAccount;
@@ -225,7 +225,7 @@ public class MainnetTransactionProcessor implements TransactionProcessor {
       }
 
       final Address senderAddress = transaction.getSender();
-      final DefaultEvmAccount sender = worldState.getOrCreate(senderAddress);
+      final EvmAccount sender = worldState.getOrCreate(senderAddress);
       validationResult =
           transactionValidator.validateForSender(transaction, sender, transactionValidationParams);
       if (!validationResult.isValid()) {
@@ -415,9 +415,8 @@ public class MainnetTransactionProcessor implements TransactionProcessor {
   }
 
   private static void clearEmptyAccounts(final WorldUpdater worldState) {
-    worldState.getTouchedAccounts().stream()
-        .filter(Account::isEmpty)
-        .forEach(a -> worldState.deleteAccount(a.getAddress()));
+    new ArrayList<>(worldState.getTouchedAccounts())
+        .stream().filter(Account::isEmpty).forEach(a -> worldState.deleteAccount(a.getAddress()));
   }
 
   private void process(final MessageFrame frame, final OperationTracer operationTracer) {
