+++ ethereum/core/src/test/java/org/hyperledger/besu/ethereum/mainnet/PrivacyBlockProcessorTest.java
@@ -29,7 +29,6 @@ import org.hyperledger.besu.ethereum.chain.Blockchain;
 import org.hyperledger.besu.ethereum.chain.TransactionLocation;
 import org.hyperledger.besu.ethereum.core.Block;
 import org.hyperledger.besu.ethereum.core.BlockDataGenerator;
-import org.hyperledger.besu.ethereum.core.DefaultEvmAccount;
 import org.hyperledger.besu.ethereum.core.Hash;
 import org.hyperledger.besu.ethereum.core.MutableAccount;
 import org.hyperledger.besu.ethereum.core.MutableWorldState;
@@ -37,6 +36,7 @@ import org.hyperledger.besu.ethereum.core.PrivateTransactionDataFixture;
 import org.hyperledger.besu.ethereum.core.TransactionReceipt;
 import org.hyperledger.besu.ethereum.core.Wei;
 import org.hyperledger.besu.ethereum.core.WorldUpdater;
+import org.hyperledger.besu.ethereum.core.WrappedEvmAccount;
 import org.hyperledger.besu.ethereum.privacy.PrivateStateRootResolver;
 import org.hyperledger.besu.ethereum.privacy.PrivateTransactionProcessor;
 import org.hyperledger.besu.ethereum.privacy.storage.PrivacyGroupHeadBlockMap;
@@ -178,10 +178,10 @@ public class PrivacyBlockProcessorTest {
   private MutableWorldState mockPrivateStateArchive() {
     final MutableWorldState mockPrivateState = mock(MutableWorldState.class);
     final WorldUpdater mockWorldUpdater = mock(WorldUpdater.class);
-    final DefaultEvmAccount mockDefaultEvmAccount = mock(DefaultEvmAccount.class);
+    final WrappedEvmAccount mockWrappedEvmAccount = mock(WrappedEvmAccount.class);
     final MutableAccount mockMutableAccount = mock(MutableAccount.class);
-    when(mockDefaultEvmAccount.getMutable()).thenReturn(mockMutableAccount);
-    when(mockWorldUpdater.createAccount(any())).thenReturn(mockDefaultEvmAccount);
+    when(mockWrappedEvmAccount.getMutable()).thenReturn(mockMutableAccount);
+    when(mockWorldUpdater.createAccount(any())).thenReturn(mockWrappedEvmAccount);
     when(mockPrivateState.updater()).thenReturn(mockWorldUpdater);
     when(mockPrivateState.rootHash()).thenReturn(Hash.ZERO);
     return mockPrivateState;
